pyproject.toml
[project]
name = "nl-to-cli-secure-agent"
version = "1.0.0"
description = "Natural Language to Windows CLI secure translation agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.7.0",
    "openai>=2.24.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]